massive simplification through offloading filesystem interface onto jmarshrossney/dcdir
